The coefficientAccessor's getLength() method is now unsigned.
This removes some sign warnings, and is overall a happier, more natural
declaration.
Unfortunately, this also means we have to declare some dicey arrays on
the Java side (unsigned ints in C++ become longs in Java). For the time
being this is just in the PM1Test, so it's okay because there we
actually have control over the image in use.

+++ stegosaurusJava/src/test/java/com/stegosaurus/steganographers/PM1Test.java
@@ -108,14 +108,14 @@ public class PM1Test extends TestWithInjection {
   public void testFakeEmbedImmutability() {
     PM1Embedder emb = embedderFactory.build(new DummyPMSequence());
     CoefficientAccessor acc = cover.getCoefficientAccessor();
-    int[] expected = new int[acc.getLength()];
+    int[] expected = new int[(int) acc.getLength()];
     for(int i = 0; i < acc.getLength(); i++) {
       expected[i] = acc.getCoefficient(i);
     }
     emb.fakeEmbed(request, SEED);
     JPEGImage other = cover.writeNew();
     acc = other.getCoefficientAccessor();
-    int[] result = new int[acc.getLength()];
+    int[] result = new int[(int) acc.getLength()];
     for(int i = 0; i < acc.getLength(); i++) {
       result[i] = acc.getCoefficient(i);
     }
